voicer/requirements.txt
livekit >= 0.17.1
livekit-agents~=0.10.1
livekit-plugins-openai~=0.10.3
python-dotenv~=1.0
loguru